requirements.txt
# Core dependencies
python-nmap>=0.7.1
scapy>=2.5.0
pyshark>=0.6.0
paramiko>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.4.2
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.4.0

# IoT specific
paho-mqtt>=1.6.1
coapthon3>=1.0.1
zeroconf>=0.38.0
# bluepy>=1.3.0  # Optional: Linux only
bleak>=0.21.1  # Windows-compatible BLE library

# Security tools
cryptography>=41.0.0
pyOpenSSL>=23.2.0
impacket>=0.10.0

# Utilities
rich>=13.5.0
typer>=0.9.0
loguru>=0.7.0
aiohttp>=3.9.0
python-socketio>=5.8.0
aiohttp-cors>=0.7.0
python-multipart>=0.0.6
websockets>=11.0.3

# New dependencies
# Add any new dependencies here 